refactor(CLI): use new `execute_safetx_flag` option to standardize
