+++ ape_safe/_cli/modules.py
@@ -2,7 +2,7 @@ import click
 from ape.cli import ConnectedProviderCommand, account_option, ape_cli_context
 from ape.types import AddressType
 
-from ape_safe._cli.click_ext import safe_argument
+from ape_safe.cli import execute_safetx_flag, safe_argument
 
 
 @click.group()
@@ -35,29 +35,29 @@ def guard(safe):
 @ape_cli_context()
 @account_option()
 @safe_argument
-@click.option("--propose", is_flag=True, default=False)
+@execute_safetx_flag()
 @click.argument("module")
-def enable(cli_ctx, safe, account, module, propose):
+def enable(cli_ctx, safe, account, module, execute_safetx):
     """
     Enable MODULE for SAFE
 
     **WARNING**: This is a potentially destructive action and may make your safe vulnerable.
     """
     module = cli_ctx.conversion_manager.convert(module, AddressType)
-    safe.modules.enable(module, submitter=account, propose=propose)
+    safe.modules.enable(module, submitter=account, propose=not execute_safetx)
 
 
 @modules.command(cls=ConnectedProviderCommand)
 @ape_cli_context()
 @account_option()
 @safe_argument
-@click.option("--propose", is_flag=True, default=False)
+@execute_safetx_flag()
 @click.argument("module")
-def disable(cli_ctx, safe, account, module, propose):
+def disable(cli_ctx, safe, account, module, execute_safetx):
     """
     Disable MODULE for SAFE
 
     **WARNING**: This is a potentially destructive action and may impact operations of your safe.
     """
     module = cli_ctx.conversion_manager.convert(module, AddressType)
-    safe.modules.disable(module, submitter=account, propose=propose)
+    safe.modules.disable(module, submitter=account, propose=not execute_safetx)
